Fix bug: get system details first, assign list details afterwards
The details of a system are cached with its full entry corresponding to the list
when they were scraped. When parsing a list, make sure that the (potentially
cached) system details are retrieved first and _then_ overwritten with the
details found on the list we are parsing, otherwise we might end up with stale
info.